requirements.txt
numpy
pandas
scikit-learn
xgboost
shap
scipy
rasterio
tqdm
pyyaml
torch